+++ tests/test_aes.py
@@ -1,5 +1,9 @@
 import os
 
+import pytest
+
+from oic.utils.aes import AEAD
+from oic.utils.aes import AESError
 from oic.utils.aes import decrypt
 from oic.utils.aes import encrypt
 
@@ -16,3 +20,50 @@ def test_encrypt_decrypt():
     encrypted_msg = encrypt(key_, msg_, 0)
     txt = decrypt(key_, encrypted_msg, 0)
     assert txt == msg_
+
+
+@pytest.fixture
+def aead_key():
+    return os.urandom(32)
+
+
+@pytest.fixture
+def aead_iv():
+    return os.urandom(16)
+
+
+@pytest.fixture
+def cleartext():
+    return b"secret sauce"
+
+
+def test_AEAD_good(aead_key, aead_iv, cleartext):
+    extra = ["some", "extra", "data"]
+    k = AEAD(aead_key, aead_iv)
+    for d in extra:
+        k.add_associated_data(d)
+    ciphertext, tag = k.encrypt_and_tag(cleartext)
+
+    # get a fresh AEAD object
+    c = AEAD(aead_key, aead_iv)
+    for d in extra:
+        c.add_associated_data(d)
+    cleartext2 = c.decrypt_and_verify(ciphertext, tag)
+    assert cleartext2 == cleartext
+
+
+def test_AEAD_bad_aad(aead_key, aead_iv, cleartext):
+    extra = ["some", "extra", "data"]
+    k = AEAD(aead_key, aead_iv)
+    for d in extra:
+        k.add_associated_data(d)
+    ciphertext, tag = k.encrypt_and_tag(cleartext)
+
+    # get a fresh AEAD object
+    c = AEAD(aead_key, aead_iv)
+    # skip one aad item, MAC is wrong now
+    for d in extra[:1]:
+        c.add_associated_data(d)
+
+    with pytest.raises(AESError):
+        c.decrypt_and_verify(ciphertext, tag)